[CARBONDATA-3889] Cleanup code in carbondata-core module
Why is this PR needed?
1.Redundant 'if' statement
2.Array access many times in for loop
3.Repeated switch branch
4.Use StringBuffer

What changes were proposed in this PR?
1.Change redundant 'if' statement
2.Use enhanced for loop
3.Merge switch branch
4.Use StringBuilder instead of StringBuffer

Does this PR introduce any user interface change?
No
Yes. (please explain the change and update document)

Is any new testcase added?
No
Yes

This closes #3868

+++ core/src/main/java/org/apache/carbondata/core/datastore/page/UnsafeDecimalColumnPage.java
@@ -51,9 +51,9 @@ public class UnsafeDecimalColumnPage extends DecimalColumnPage {
       int size = pageSize * 3;
       memoryBlock = UnsafeMemoryManager.allocateMemoryWithRetry(taskId, size);
     } else if (DataTypes.isDecimal(columnPageEncoderMeta.getStoreDataType())) {
-      memoryBlock = UnsafeMemoryManager.allocateMemoryWithRetry(taskId, (long) (capacity));
+      memoryBlock = UnsafeMemoryManager.allocateMemoryWithRetry(taskId, capacity);
     } else if (columnPageEncoderMeta.getStoreDataType() == DataTypes.BYTE_ARRAY) {
-      memoryBlock = UnsafeMemoryManager.allocateMemoryWithRetry(taskId, (long) (capacity));
+      memoryBlock = UnsafeMemoryManager.allocateMemoryWithRetry(taskId, capacity);
     } else {
       throw new UnsupportedOperationException(
           "invalid data type: " + columnPageEncoderMeta.getStoreDataType());
